lib_conference_spider/requirements.txt
requests
beautifulsoup4
pandas
deepl
langchain-community
langchain-core
python-dotenv
selenium
webdriver-manager